pyproject.toml
[project]
name = "aitrace-datasets"
version = "0.1.0"
description = "AITrace Datasets - Dataset management tool with Airtable-inspired UI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.1",
    "python-multipart>=0.0.12",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "cloud-sql-python-connector[asyncpg]>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "isort>=5.13.0",
    "mypy>=1.13.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.hatch.build.targets.wheel]
packages = ["src/aitrace"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
